examples/python/requirements.txt
aws-cdk.core
cdk-metaflow